requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-login==0.5.0
scapy==2.4.5
psutil==5.8.0
sqlalchemy==1.4.23
cryptography==3.4.7
requests==2.26.0
python-dotenv==0.19.0
pytest==6.2.5
black==21.7b0
pylint==2.9.6
aiohttp==3.7.4
asyncio==3.4.3
numpy==1.21.2
pandas==1.3.2
prometheus-client==0.11.0
PyYAML==5.4.1
redis==3.5.3
elasticsearch==7.14.0
kafka-python==2.0.2
scikit-learn==0.24.2
scipy==1.7.1
python-jose==3.3.0
gunicorn==20.1.0
eventlet==0.33.0
gevent==21.8.0 